# shippingCloud.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "shippingManagement",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/order/company/{id}/order": {
      "post": {
        "summary": "承运商接单创建订单",
        "x-apifox-folder": "order_service",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "order_service"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "2",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJqdGkiOiJhNTc3Mjc2MDEwOTY0YzM4YWIwNzJiMTdkYmE0NDg3OCIsInN1YiI6IjIiLCJpc3MiOiJ5amwiLCJpYXQiOjE2ODY5NjYxNjAsImV4cCI6MTY4Njk2OTc2MH0.9A8VNfYd_hFWXryWlCnr16j-NuNOj69b1hwHvYKrZrk",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apifox-orders": [
                  "01GZ6HQNX5P1T7Z5A23F13RF12"
                ],
                "required": [
                  "01GZ6HQNX5P1T7Z5A23F13RF12"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": {
                "goodsId": "3",
                "driverId": "2",
                "carId": "京B333"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/2565994/apis/api-78950183-run"
      }
    },
    "/order/company/{id}/order/{orderId}/{status}/{now_addr}": {
      "put": {
        "summary": "更新订单状态和所在地址",
        "x-apifox-folder": "order_service",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "order_service"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "2",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "now_addr",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "",
            "required": false,
            "example": "运输中",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "now_addr",
            "in": "query",
            "description": "",
            "required": false,
            "example": "南京",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJqdGkiOiJkNTJlNTIzNDNjNWY0ZDY1YWU4ZjIzOGRhMmE5ODc3MiIsInN1YiI6IjIiLCJpc3MiOiJ5amwiLCJpYXQiOjE2ODI3NzIxODgsImV4cCI6MTY4Mjc3NTc4OH0.m7q9Qm4JhJsc_3CEBU9gRzi_YQjzxG2ZfB72OTRMtV4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/2565994/apis/api-78950336-run"
      }
    },
    "/order/company/{id}/order/{orderId}/statue/signed": {
      "put": {
        "summary": "更新订单状态为已签收",
        "x-apifox-folder": "order_service",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "order_service"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJqdGkiOiIzYzNlMzBlYjA0MzE0YWY3YjQ2ZjA2YTc1ZGZkMDkzYiIsInN1YiI6IjEiLCJpc3MiOiJ5amwiLCJpYXQiOjE2ODI3NzM4NjIsImV4cCI6MTY4Mjc3NzQ2Mn0.mZMcskX5T0oqQscidp0lXCA2U6j22QQCwFIr5rIqDFQ",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-ignore-properties": [],
                  "x-apifox-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/2565994/apis/api-78950373-run"
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:8083",
      "description": "开发环境"
    }
  ]
}